+++ lib/src/modules/geo/geomerty/spatial_reference_system.dart
@@ -1,0 +1,10 @@
+part of backendless_sdk;
+
+enum SpatialReferenceSystem {
+  CARTESIAN,
+  PULKOVO_1995,
+  WGS84,
+  WGS84_PSEUDO_MERCATOR,
+  WGS84_WORLD_MERCATOR,
+  UNKNOWN
+}