+++ lib/src/modules/user_service/user_service.dart
@@ -63,32 +63,45 @@ class BackendlessUserService {
       "Backendless.UserService.setUserToken",
       <String, dynamic>{"userToken": userToken});
 
+  Future<BackendlessUser> loginAsGuest([bool stayLoggedIn]) =>
+      _channel.invokeMethod("Backendless.UserService.loginAsGuest",
+          <String, dynamic>{"stayLoggedIn": stayLoggedIn});
+
   Future<BackendlessUser> loginWithFacebook(String accessToken,
-          {Map<String, String> fieldsMapping, bool stayLoggedIn}) =>
+          {Map<String, String> fieldsMapping,
+          bool stayLoggedIn,
+          BackendlessUser guestUser}) =>
       _channel.invokeMethod(
           "Backendless.UserService.loginWithFacebook", <String, dynamic>{
         "accessToken": accessToken,
         "fieldsMapping": fieldsMapping,
-        "stayLoggedIn": stayLoggedIn
+        "stayLoggedIn": stayLoggedIn,
+        "guestUser": guestUser
       });
 
   Future<BackendlessUser> loginWithTwitter(
           String authToken, String authTokenSecret,
-          {Map<String, String> fieldsMapping, bool stayLoggedIn}) =>
+          {Map<String, String> fieldsMapping,
+          bool stayLoggedIn,
+          BackendlessUser guestUser}) =>
       _channel.invokeMethod(
           "Backendless.UserService.loginWithTwitter", <String, dynamic>{
         "authToken": authToken,
         "authTokenSecret": authTokenSecret,
         "fieldsMapping": fieldsMapping,
-        "stayLoggedIn": stayLoggedIn
+        "stayLoggedIn": stayLoggedIn,
+        "guestUser": guestUser
       });
 
   Future<BackendlessUser> loginWithGoogle(String accessToken,
-          {Map<String, String> fieldsMapping, bool stayLoggedIn}) =>
+          {Map<String, String> fieldsMapping,
+          bool stayLoggedIn,
+          BackendlessUser guestUser}) =>
       _channel.invokeMethod(
           "Backendless.UserService.loginWithGoogle", <String, dynamic>{
         "accessToken": accessToken,
         "fieldsMapping": fieldsMapping,
-        "stayLoggedIn": stayLoggedIn
+        "stayLoggedIn": stayLoggedIn,
+        "guestUser": guestUser
       });
 }