+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+* Support for geometry data types
+* Guest login with social account
+* Custom classes caching
+
 ## 1.0.1
 
 * Fix crash during publish() call